requirements.txt
ConfigSpace==0.7.1
cryptography>=42.0.4
debugpy
decorator 
exceptiongroup 
executing
gym
jax[cpu]
jedi
matplotlib
matplotlib-inline 
more-itertools==10.1.0
nest-asyncio 
numpy
numpyro
packaging 
pandas
paramiko>=3.4.0
parso 
pexpect 
pickleshare 
platformdirs 
prompt-toolkit 
psutil 
psycopg2-binary==2.9.9
ptyprocess 
pure-eval 
pycparser==2.21
Pygments 
PyNaCl==1.5.0
pyparsing==3.1.1
python-dateutil 
pyzmq 
scipy==1.11.4
smac==1.4.0
six 
stack-data
tornado 
traitlets 
typing_extensions 
wcwidth 
zipp 
pyyaml
torch
botorch
pyDOE
pygmo